Move system tests to test_mailmerge.py
