+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""Dummy file needed for pylint to discover unit tests."""
