+++ tox.ini
@@ -15,5 +15,5 @@ commands =
     check-manifest
     pycodestyle mailmerge tests setup.py
     pydocstyle mailmerge tests setup.py
-    pylint mailmerge setup.py cd.py test_mailmerge.py test_message_template.py test_sendmail_client.py utils.py
+    pylint mailmerge setup.py tests
     pytest --verbose --cov mailmerge
